facebook_api_to_json/requirements.txt
facebook-business>=19.0.0
python-dotenv>=1.0.0
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
SQLAlchemy>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2 